requirements.txt
Flask
requests
python-whois
dnspython
python-Levenshtein